+++ cerulean/credential.py
@@ -1,17 +1,16 @@
 from abc import ABC
+from typing import Optional
 
 
 class Credential(ABC):
     """A credential for connecting to remote machines.
 
-    Credentials don't have much in common other than a username, \
-    which is best modelled as a public attribute. So this interface \
-    is empty, and only here to provide a generic type to represent \
-    any credential in the API.
+    Credentials don't have much in common other than a username, which is best
+    modelled as a public attribute. So this interface is empty, and only here
+    to provide a generic type to represent any credential in the API.
 
     Attributes:
         username: The name of the user to connect as.
-
     """
     pass
 
@@ -22,16 +21,13 @@ class PasswordCredential(Credential):
     Attributes:
         username: The name of the user to connect as
         password: The password to authenticate with
-
     """
-
     def __init__(self, username: str, password: str) -> None:
         """Create a PasswordCredential.
 
         Args:
             username: The name of the user to connect as
             password: The password to authenticate with
-
         """
         self.username = username
         self.password = password
@@ -44,18 +40,16 @@ class PubKeyCredential(Credential):
         username: The name of the user to connect as
         public_key: The (local) path to a key file
         passphrase: The passphrase to decrypt the key with; optional.
-
     """
-
-    def __init__(self, username: str, public_key: str,
-                 passphrase: str = None) -> None:
+    def __init__(
+            self, username: str, public_key: str, passphrase: Optional[str] = None
+            ) -> None:
         """Create a PubKeyCredential.
 
         Args:
             username: The name of the user to connect as
-            pub_key: The (local) path to a key file
-            passphrase The passphrase to decrypt the key with; optional.
-
+            public_key: The (local) path to a key file
+            passphrase: The passphrase to decrypt the key with; optional.
         """
         self.username = username
         self.public_key = public_key