+++ cerulean/direct_gnu_scheduler.py
@@ -13,15 +13,14 @@ logger = logging.getLogger(__name__)
 class DirectGnuScheduler(Scheduler):
     """A scheduler that runs processes directly on a GNU system.
 
-    This scheduler does not have a queue, instead it launches each \
-    job immediately as a process, and uses ps and kill to manage it. \
+    This scheduler does not have a queue, instead it launches each job
+    immediately as a process, and uses ps and kill to manage it.
 
-    This should work fine on any normal GNU/Linux system, but in some
-    cases you may need an extra command to make bash, ps and/or kill
-    available (e.g. setting a PATH). If so, you can specify prefix, and
-    it will be prepended onto these commands. Note that this is a
-    simple string concatenation, so you may need a semicolon at the
-    end depending your exact prefix command.
+    This should work fine on any normal GNU/Linux system, but in some cases you
+    may need an extra command to make bash, ps and/or kill available (e.g.
+    setting a PATH). If so, you can specify prefix, and it will be prepended
+    onto these commands. Note that this is a simple string concatenation, so
+    you may need a semicolon at the end depending your exact prefix command.
     """
     def __init__(self, terminal: Terminal, prefix: str = '') -> None:
         """Create a DirectGnuScheduler.
@@ -29,7 +28,6 @@ class DirectGnuScheduler(Scheduler):
         Args:
             terminal: The terminal to execute on.
             prefix: A string to prefix the shell commands with.
-
         """
         self.__terminal = terminal
         self.__prefix = prefix
@@ -40,34 +38,45 @@ class DirectGnuScheduler(Scheduler):
 
         if job_description.mpi_processes_per_node is not None:
             raise RuntimeError(
-                "mpi_processes_per_node is not supported by DirectGnuScheduler, "
-                "because we cannot inject this into the MPI configuration in an environment "
-                "without a scheduler. You should call mpirun with an appropriate parameter "
-                "instead.")
+                'mpi_processes_per_node is not supported by DirectGnuScheduler,'
+                ' because we cannot inject this into the MPI configuration in an'
+                ' environment without a scheduler. You should call mpirun with an'
+                ' appropriate parameter instead.')
 
         job_script = ''
+
         for name, value in job_description.environment.items():
             job_script += "export {}='{}'\n".format(name, value)
+
         if job_description.working_directory is not None:
             job_script += 'cd {}\n'.format(job_description.working_directory)
+
         job_script += 'exit_code_file=$(mktemp)\n'
         job_script += "(\n"
-        if job_description.time_reserved is not None:
-            job_script += "ulimit -t {}\n".format(
-                job_description.time_reserved)
-        escaped_command = job_description.command.replace("'", "'\\\''")    # type: ignore
+
+        escaped_command = job_description.command.replace(
+                "'", "'\\\''")    # type: ignore
         escaped_args = map(lambda s: s.replace("'", "'\\\''"),
                            job_description.arguments)
+
         job_script += "bash -c '"
+
+        if job_description.time_reserved is not None:
+            job_script += "timeout {} ".format(job_description.time_reserved)
+
         job_script += '{}'.format(escaped_command)
         job_script += ' {}'.format(' '.join(escaped_args))
+
         if job_description.stdout_file is not None:
             job_script += ' >{}'.format(job_description.stdout_file)
+
         if job_description.stderr_file is not None:
             job_script += ' 2>{}'.format(job_description.stderr_file)
+
         job_script += "' ; "
         job_script += 'echo $? >$exit_code_file'
         job_script += ')'
+
         if job_description.system_out_file is not None:
             job_script += ' >{}'.format(job_description.system_out_file)
         else:
@@ -77,14 +86,16 @@ class DirectGnuScheduler(Scheduler):
             job_script += ' 2>{}'.format(job_description.system_err_file)
         else:
             job_script += ' 2>/dev/null'
+
         job_script += ' &\n'
+
         job_script += 'echo -n $! $exit_code_file\n'
         job_script += 'disown\n'
 
         logger.debug('Job script: %s', job_script)
         command = self.__prefix + ' bash'
-        _, output, error = self.__terminal.run(10.0, command, [],
-                                                       job_script)
+        _, output, error = self.__terminal.run(
+                10.0, command, [], job_script)
 
         return output