+++ tests/container-test/endless-job.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Sleep forever
+while true ; do
+    sleep 10
+done
+